% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcontax-package.R
\docType{package}
\name{microcontax-package}
\alias{microcontax-package}
\alias{microcontax}
\title{The ConTax data package}
\usage{
microcontax()
}
\description{
The consensus taxonomy for prokaryotes is a package of data sets designed to be the best
possible for training taxonomic classifiers based on 16S rRNA sequence data.
}
\details{
\tabular{ll}{
Package: \tab microcontax\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2020-06-06\cr
License: \tab GPL-2\cr
}
}
\seealso{
\code{\link[microseq:microseq-package]{microseq}}
}
\author{
Hilde Vinje, Kristian Liland, Lars Snipen.\cr
Maintainer: Lars Snipen <lars.snipen@nmbu.no>
}
\keyword{package}
